MINT-847 remove window-specific code from IPFS (#471)
